requirements.txt
# Core dependencies
groq==0.4.1
torch==2.2.0
transformers==4.37.2
numpy==1.24.3
python-dotenv==1.0.0
aiohttp==3.9.1
typing-extensions==4.12.2
chromadb==0.4.22
sentence-transformers==2.3.0
spacy==3.8.0

# Visualization dependencies
matplotlib==3.8.1
networkx==3.2.1
wordcloud==1.9.2
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
pyvis==0.3.2  # For memory graph visualization

# Additional dependencies
neo4j
altair==5.4.1
pydeck==0.9.1
pillow==11.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# LLM dependencies
openai==1.60.1
langchain_groq==0.2.3
langchain==0.3.15

# Performance and optimization
tqdm==4.66.1  # For progress bars
ujson==5.8.0  # Faster JSON processing 